# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Devot API",
        "version": "0.1"
    },
    "paths": {
        "/api/budget/summary": {
            "post": {
                "tags": [
                    "Budget Summary"
                ],
                "summary": "Retrieves a summary of expenses",
                "description": "Retrieves a summary of expenses based on a validated input. Returns the total amount spent and expense data.",
                "operationId": "bcbd56ddc3f4ea4b33145dbe7653e00f",
                "requestBody": {
                    "description": "The request body containing the validated input.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetSummaryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_spent": {
                                                    "description": "The total amount spent on expenses.",
                                                    "type": "number",
                                                    "example": 1000.5
                                                },
                                                "data": {
                                                    "description": "The array of expenses.",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Expense"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Retrieves a collection of categories for the authenticated user",
                "description": "Retrieves a collection of categories for the authenticated user.",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Successful response with a collection of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/categories": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category with the given details.",
                "operationId": "store",
                "requestBody": {
                    "description": "Category details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Category Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Category name",
                        "type": "string",
                        "example": "Car expenses"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Category"
                }
            },
            "Expense": {
                "title": "Expense",
                "description": "Expense",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Expense name",
                        "type": "string",
                        "example": "Car expenses"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Expence amount",
                        "type": "number",
                        "format": "float",
                        "example": "124.33"
                    },
                    "description": {
                        "title": "description",
                        "description": "Expense description",
                        "type": "string",
                        "example": "Car expenses for March."
                    },
                    "category": {
                        "title": "id",
                        "description": "Expense category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "date": {
                        "title": "id",
                        "description": "Expense category",
                        "type": "string",
                        "example": "27.03.2023"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Expense"
                }
            },
            "BudgetSummaryRequest": {
                "title": "Login Request",
                "description": "Login Request request body data",
                "properties": {
                    "start_date": {
                        "title": "email",
                        "description": "List data that is created after this date",
                        "type": "string",
                        "default": "Start of the current month",
                        "example": "21.03.2022"
                    },
                    "end_date": {
                        "title": "end date",
                        "description": "List data that is created before this date",
                        "type": "string",
                        "default": "Today",
                        "example": "21.03.2023"
                    },
                    "categores": {
                        "title": "categories",
                        "description": "List of categories to apply",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": "[1, 4, 55]"
                    }
                },
                "type": "object"
            },
...
            "StoreCategoryRequest": {
                "title": "Store Expense Request",
                "description": "Store Expense request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Category name",
                        "type": "string",
                        "example": "Car expenses"
                    }
                },
                "type": "object"
            },
...
            "CategoryResource": {
                "title": "CategoryResource",
                "description": "Category resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CategoryResource"
                }
            },
...
        }
    }
}